executor: Single K8s Job (#53311)
Part of #50601 (not the actual full implementation as corners were cut).

## Changes

- Created a specific package to handle workspace files
  - Makes it easier to have same functionality across "modes" 
  - Moved `workspace.FileStore` to `files.Store`
- Created env vars to configuring the single job
- The volume can be a dynamically created PVC or an `emptyDir` with a
size limit
- Reading of container logs is now kicked off during the pod watching. 
- This makes it easier to have shared functionality between the
different modes
  - Can kick off logs based on the container status vs waiting
- For the single job, a the job token is put into a secret that is
created and deleted for the job
- Refactored the Command Logger into a separate package to avoid
cyclical imports
  - Will also make it easier to use in a custom image

### Important Changes

- `enterprise/cmd/executor/internal/worker/files/files.go`
- `enterprise/cmd/executor/internal/worker/runtime/kubernetes.go`
- `enterprise/cmd/executor/internal/worker/runner/kubernetes.go`
- `enterprise/cmd/executor/internal/worker/command/kubernetes.go`

### The Good

- Handles step skipping
- Handle workspace files
- Handles env outputs
- Handles previous step outputs
- Hidden behind an environment variables. So can be turned off and on
- Refactoring has been done to make it easier to extract functionality
into another image to do this instead of hardcoding into the init
containers

### The Bad

- Using `batcheshlper` to run the setup `initContainer` and the `main`
container
    - This image comes with `git`, so it works out well for the cloning
- This image is already (probably) already being pulled in, so one less
thing for a user to bring in
- There is some hardcoded nonsense (e.g. step skipping) 
- Cannot handle large workspace files (e.g. binaries)

## Test plan

- Updated existing tests
- Added new tests
- Ran tests with `docker` runtime
- Ran tests with `kubernetes` runtime
  - Tested existing functionality
  - Tested single job functionality

+++ enterprise/cmd/executor/internal/worker/workspace/kubernetes.go
@@ -6,27 +6,35 @@ import (
 	"os"
 	"path/filepath"
 
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/executor/internal/worker/cmdlogger"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/executor/internal/worker/command"
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/executor/internal/worker/files"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/executor/types"
 )
 
 type kubernetesWorkspace struct {
 	scriptFilenames []string
 	workspaceDir    string
-	logger          command.Logger
+	logger          cmdlogger.Logger
 }
 
 // NewKubernetesWorkspace creates a new workspace for a job.
 func NewKubernetesWorkspace(
 	ctx context.Context,
-	filesStore FilesStore,
+	filesStore files.Store,
 	job types.Job,
 	cmd command.Command,
-	logger command.Logger,
+	logger cmdlogger.Logger,
 	cloneOpts CloneOptions,
 	mountPath string,
+	singleJob bool,
 	operations *command.Operations,
 ) (Workspace, error) {
+	// TODO switch to the single job in 5.2
+	if singleJob {
+		return &kubernetesWorkspace{logger: logger}, nil
+	}
+
 	workspaceDir := filepath.Join(mountPath, fmt.Sprintf("job-%d", job.ID))
 
 	if err := os.MkdirAll(workspaceDir, os.ModePerm); err != nil {
@@ -79,8 +87,10 @@ func (w kubernetesWorkspace) Remove(ctx context.Context, keepWorkspace bool) {
 		return
 	}
 
-	fmt.Fprintf(handle, "Removing %s\n", w.workspaceDir)
-	if rmErr := os.RemoveAll(w.workspaceDir); rmErr != nil {
-		fmt.Fprintf(handle, "Operation failed: %s\n", rmErr.Error())
+	if w.workspaceDir != "" {
+		fmt.Fprintf(handle, "Removing %s\n", w.workspaceDir)
+		if rmErr := os.RemoveAll(w.workspaceDir); rmErr != nil {
+			fmt.Fprintf(handle, "Operation failed: %s\n", rmErr.Error())
+		}
 	}
 }
